fix(hailo): force FLOAT32 outputs + drop /255 input pre-norm
Two refactor regressions surfaced by the bench:

1. Output dtype: pre-compiled HEFs natively emit varying dtypes
   (NMS YOLO is FLOAT32; ResNet50 softmax is UINT8 quantized). The
   old API set format_type=FLOAT32 and HailoRT dequantized for us;
   the new InferModel API needs an explicit set_format_type call,
   BEFORE configure(). Now done in HailoModel._ensure_configured for
   all outputs — caller always reads FLOAT32.

2. Input dtype: HailoClassifier was passing FLOAT32 [0,1]-normalized
   pixels, relying on the old API's auto-quantization. Pre-compiled
   HEFs from /usr/share/hailo-models bake the normalization into the
   graph (per playbook §5.1), so passing raw UINT8 0-255 is correct
   and matches HailoDetector's contract. Drop the /255 step.

Bench verified on Pi 5 + Hailo-8L: DET 58.9 FPS isolated → 45.5 FPS
co-scheduled with CLS; CLS 47.7 → 44.2 FPS. ~6ms scheduler overhead
per iter. Both well above the pipeline's 5 FPS sub-stream target.

+++ tools/bench_hailo_multimodel.py
@@ -89,8 +89,10 @@ def main():
     det = eng.acquire_model(args.det)
     cls = eng.acquire_model(args.cls)
 
+    # Both detector and classifier HEFs from /usr/share/hailo-models bake
+    # any required normalization into the graph; pass raw UINT8 0-255 pixels.
     det_in = np.zeros(det.input_shape(), dtype=np.uint8)
-    cls_in = np.zeros(cls.input_shape(), dtype=np.float32)
+    cls_in = np.zeros(cls.input_shape(), dtype=np.uint8)
 
     print(f"\n=== Single-model isolated (warmup={args.warmup}, n={args.n}) ===")
     for label, m, x in (("DET", det, det_in), ("CLS", cls, cls_in)):
